requirements.txt
# openai==0.28
openai==1.99.9
pytest==8.4.1
python-dotenv==1.1.1
streamlit==1.44.1
